AssessmentLO3
Programme is designed to sort an Arraylist of integers through a user interface. The programme makes use of the bubble sort and merge sort sorting algorithms, additionally measuring the performance of each algorithm in terms of execution run time as well as the number of swaps and comparisons the algorithm had to make to sort the Data.